refact: moved security configs to folder 'config.security'
